Module_5/langchain_llamaindex/requirements.txt
langchain
langchain-community
langchain-google-genai

llama-index
llama-index-llms-google-genai
llama-index-embeddings-huggingface

faiss-cpu
sentence-transformers
python-dotenv
huggingface-hub